refactor(example): restructure example project folders
- Split into basic_example, nullable_example, constant_example, and multi_env_example
- Split multi_env_example into merge_in_one_file and separated_files
