+++ examples/envied_example/lib/constant_example/constant_case_env.dart
