+++ examples/envied_example/lib/multi_env_example/merge_in_one_file/envs.env.dart
@@ -1,8 +1,8 @@
 // lib/env.dart
+import 'package:example/multi_env_example/app_env_fields.dart';
+
 import 'package:envied/envied.dart';
 import 'package:example/example_enum.dart';
-
-import 'app_env_fields.dart';
 
 part 'envs.env.g.dart';
