+++ examples/envied_example/lib/multi_env_example/separated_files/app_env.dart
@@ -1,4 +1,5 @@
-import 'app_env_fields.dart';
+import 'package:example/multi_env_example/app_env_fields.dart';
+
 import 'debug_env.dart';
 import 'release_env.dart';
